goose: Restrict focus to revdeps testing
Goose is an agentic framework for writing AI workflows using
natural language recipes. It was experimented with initially
in this repository for RHEL agentic workflows. These days it
is only used for reverse dependency testing.

Given the narrowed scope of goose, and in preparation for getting rid
of goose entirely, this commit drops all recipes but the one that's
still being used.